refactor: GPT Controller 및 Service 분리
